+++ src/components/Action/index.ts
@@ -1,0 +1,2 @@
+export * from "./Action.component";
+export type * from "./Action.types";