+++ src/components/Icon/__generated__/custom/index.ts
@@ -3,7 +3,7 @@ export { AppsFill } from "./AppsFill.icon";
 export { BuildingBlocks } from "./BuildingBlocks.icon";
 export { BuildingBlocksFill } from "./BuildingBlocksFill.icon";
 export { CatRobot } from "./CatRobot.icon";
-export { Delete } from "./Delete.icon";
+export { CircleDashed } from "./CircleDashed.icon";
 export { Figma } from "./Figma.icon";
 export { FigmaFill } from "./FigmaFill.icon";
 export { Fire } from "./Fire.icon";