+++ src/utils/component.utils.ts
@@ -1,0 +1,69 @@
+import type { TestIdProps } from "@types";
+import type { FC } from "react";
+
+type Component<TComponentProps extends Record<string, unknown>> =
+  TComponentProps extends infer TBaseProps & TestIdProps
+    ?
+        | FC<TBaseProps & TestIdProps>
+        | (FC<TComponentProps> & { defaultProps: Required<TBaseProps> })
+    :
+        | FC<TComponentProps>
+        | (FC<TComponentProps> & { defaultProps: Required<TComponentProps> });
+
+type ComponentModule<
+  TComponentName extends string,
+  TComponentProps extends Record<string, unknown>
+> = Record<string, unknown> & {
+  [key in TComponentName]: Component<TComponentProps>;
+};
+
+type ComponentDefaultModule<TComponentProps extends Record<string, unknown>> = {
+  default: Component<TComponentProps>;
+};
+
+/**
+ * Loads a component from a specified path and returns it as a Promise.
+ *
+ * @remarks used for lazy-loading a component dynamically
+ *
+ * @param {string} componentName - The name of the component to load.
+ * @param {string} [componentPath] - The path to the component module.
+ * @return {Promise<ComponentDefaultModule<TComponentProps>>} A Promise that resolves to the loaded component module.
+ *
+ * @example
+ * ```tsx
+ * import { lazy, useReducer } from "react";
+ * import { loadComponent } from "@utils";
+ *
+ * const LazyLoadedBadge = lazy(loadComponent("Badge", import("../Badge")));
+ *
+ * export const MyComponent = () => {
+ *  const [state, toggleState] = useReducer((state) => !state, false);
+ *  return (
+ *    <div>
+ *      <span>Label</span>
+ *      {state && <LazyLoadedBadge label="Badge"/>}
+ *    <div/>
+ *  )
+ * };
+ * ```
+ */
+export const loadComponent =
+  <
+    TComponentName extends string,
+    TComponentProps extends Record<string, unknown>
+  >(
+    componentName: TComponentName,
+    modulePromise: Promise<ComponentModule<TComponentName, TComponentProps>>
+  ) =>
+  async (): Promise<ComponentDefaultModule<TComponentProps>> => {
+    try {
+      const module = await modulePromise;
+      if (!(componentName in module)) {
+        throw new Error(`Component ${componentName} not found`);
+      }
+      return { default: module[componentName] };
+    } catch (e) {
+      throw new Error(`Failed to load component ${componentName}: ${e}`);
+    }
+  };