+++ src/utils/index.ts
@@ -4,3 +4,4 @@ export * from "./log.utils.ts";
 export * from "./number.utils";
 export * from "./props.utils.ts";
 export * from "./string.utils.ts";
+export * from "./component.utils.ts";